requirements.txt
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
discord.py==2.3.2 